Python: inviwopy cleanup and fixes
