+++ modules/python3/bindings/include/inviwopy/inviwopy.h
@@ -29,11 +29,6 @@
 
 #pragma once
 
-#include <warn/push>
-#include <warn/ignore/shadow>
-#include <pybind11/pybind11.h>
-#include <warn/pop>
-
 namespace pybind11 {
