+++ modules/python3/bindings/include/inviwopy/pybuffer.h
@@ -32,7 +32,6 @@
 #include <warn/push>
 #include <warn/ignore/shadow>
 #include <pybind11/pybind11.h>
-#include <pybind11/numpy.h>
 #include <warn/pop>
 
 namespace inviwo {
