+++ modules/python3/bindings/include/inviwopy/pycompositeproperties.h
@@ -2,7 +2,7 @@
  *
  * Inviwo - Interactive Visualization Workshop
  *
- * Copyright (c) 2018-2021 Inviwo Foundation
+ * Copyright (c) 2017-2021 Inviwo Foundation
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -35,7 +35,7 @@
 #include <warn/pop>
 
 namespace inviwo {
-
-void exposeTFPrimitiveSet(pybind11::module& m);
+    
+void exposeCompositeProperties(pybind11::module& m);
 
 }  // namespace inviwo
