+++ modules/python3/bindings/include/inviwopy/pyglmports.h
@@ -29,8 +29,6 @@
 
 #pragma once
 
-#include <inviwo/core/util/glm.h>
-
 #include <warn/push>
 #include <warn/ignore/shadow>
 #include <pybind11/pybind11.h>
