+++ modules/python3/bindings/include/inviwopy/pyimage.h
@@ -29,12 +29,12 @@
 
 #pragma once
 
-#include <inviwo/core/datastructures/image/imagetypes.h>
-
 #include <warn/push>
 #include <warn/ignore/shadow>
 #include <pybind11/pybind11.h>
 #include <warn/pop>
+
+#include <inviwo/core/datastructures/image/imagetypes.h>
 
 PYBIND11_MAKE_OPAQUE(inviwo::SwizzleMask)
