+++ modules/python3/bindings/include/inviwopy/pynetwork.h
@@ -34,8 +34,6 @@
 #include <pybind11/pybind11.h>
 #include <warn/pop>
 
-#include <inviwo/core/ports/outport.h>
-
 namespace inviwo {
 
 void exposeNetwork(pybind11::module& m);
