+++ modules/python3/bindings/include/inviwopy/pyvolume.h
@@ -37,4 +37,5 @@
 namespace inviwo {
 
 void exposeVolume(pybind11::module& m);
+
 }
