+++ modules/python3/bindings/src/pyimage.cpp
@@ -28,23 +28,18 @@
  *********************************************************************************/
 
 #include <inviwopy/pyimage.h>
+#include <inviwopy/pyglmtypes.h>
 
 #include <inviwo/core/util/formatdispatching.h>
 #include <inviwo/core/datastructures/image/layer.h>
 #include <inviwo/core/datastructures/image/image.h>
 #include <inviwo/core/datastructures/image/layerram.h>
 #include <inviwo/core/datastructures/image/layerramprecision.h>
-
 #include <inviwo/core/io/datawriterfactory.h>
 #include <inviwo/core/util/filesystem.h>
 #include <inviwo/core/util/stringconversion.h>
 #include <inviwo/core/common/inviwoapplication.h>
-
 #include <inviwo/core/ports/imageport.h>
-
-#include <inviwopy/inviwopy.h>
-#include <inviwopy/pynetwork.h>
-#include <inviwopy/pyglmtypes.h>
 #include <modules/python3/pybindutils.h>
 #include <modules/python3/pyportutils.h>
