+++ modules/python3/bindings/src/pypropertyowner.cpp
@@ -28,12 +28,10 @@
  *********************************************************************************/
 
 #include <inviwopy/pypropertyowner.h>
+#include <inviwopy/vectoridentifierwrapper.h>
+
 #include <inviwo/core/properties/propertyowner.h>
 #include <inviwo/core/processors/processor.h>
-
-#include <inviwopy/inviwopy.h>
-#include <inviwopy/vectoridentifierwrapper.h>
-#include <inviwopy/pyproperties.h>
 
 #include <pybind11/detail/common.h>
 
@@ -45,16 +43,16 @@ void exposePropertyOwner(pybind11::module& m) {
     using PropertyVecWrapper = VectorIdentifierWrapper<std::vector<Property*>>;
     exposeVectorIdentifierWrapper<std::vector<Property*>>(m, "PropertyVecWrapper");
 
-    py::class_<PropertyOwner>(m, "PropertyOwner")
+    py::class_<PropertyOwner>(m, "PropertyOwner", py::multiple_inheritance{}, py::dynamic_attr{})
         .def(
             "__getattr__",
             [](PropertyOwner& po, const std::string& key) {
                 if (auto prop = po.getPropertyByIdentifier(key)) {
                     return prop;
                 } else {
-                    throw py::attribute_error{"PropertyOwner (" + po.getIdentifier() +
-                                              ") does not have a property with identifier: '" +
-                                              key + "'"};
+                    throw py::attribute_error{fmt::format(
+                        "PropertyOwner ({}) does not have a property with identifier: '{}'",
+                        po.getIdentifier(), key)};
                 }
             },
             py::return_value_policy::reference)
