+++ modules/python3/bindings/src/pypropertytypehook.cpp
@@ -2,7 +2,7 @@
  *
  * Inviwo - Interactive Visualization Workshop
  *
- * Copyright (c) 2018-2021 Inviwo Foundation
+ * Copyright (c) 2017-2021 Inviwo Foundation
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -27,15 +27,4 @@
  *
  *********************************************************************************/
 
-#pragma once
-
-#include <warn/push>
-#include <warn/ignore/shadow>
-#include <pybind11/pybind11.h>
-#include <warn/pop>
-
-namespace inviwo {
-
-void exposeTFPrimitiveSet(pybind11::module& m);
-
-}  // namespace inviwo
+#include <inviwopy/pypropertytypehook.h>
