+++ modules/python3/bindings/src/pyvolume.cpp
@@ -28,32 +28,24 @@
  *********************************************************************************/
 
 #include <inviwopy/pyvolume.h>
-
-#include <inviwo/core/util/formatdispatching.h>
-
-#include <inviwopy/inviwopy.h>
-#include <inviwopy/pynetwork.h>
 #include <inviwopy/pyglmtypes.h>
 #include <inviwopy/pyimage.h>  // for the opaque swizzlemask
-#include <modules/python3/pybindutils.h>
-#include <modules/python3/pyportutils.h>
-#include <modules/python3/volumepy.h>
 
-#include <warn/push>
-#include <warn/ignore/shadow>
-#include <pybind11/pybind11.h>
-#include <pybind11/numpy.h>
-#include <pybind11/stl_bind.h>
-#include <warn/pop>
-
+#include <inviwo/core/util/formatdispatching.h>
 #include <inviwo/core/datastructures/volume/volume.h>
 #include <inviwo/core/datastructures/volume/volumeram.h>
 #include <inviwo/core/datastructures/volume/volumeramprecision.h>
 #include <inviwo/core/ports/volumeport.h>
+#include <inviwo/core/datastructures/unitsystem.h>
+#include <modules/python3/pybindutils.h>
+#include <modules/python3/pyportutils.h>
+#include <modules/python3/volumepy.h>
+
+#include <pybind11/numpy.h>
+#include <pybind11/stl_bind.h>
 
 #include <fmt/format.h>
 #include <fmt/ostream.h>
-#include <units/units.hpp>
 
 PYBIND11_MAKE_OPAQUE(inviwo::VolumeSequence)
